moved displayAllGamesTest to the gameList test file. displayAllGames implemented
