+++ src/main/java/GameList.java
@@ -62,7 +62,17 @@ public class GameList {
     }
 
     public void displayAllGames(){
+        System.out.println(name + ":\n");
 
+        if (getGameCount()==0){
+            System.out.println("This list is empty");
+        }
+        else {
+            for (int i = 0; i < getGameCount(); i++) {
+                gameList.get(i).displayGame();
+                //System.out.println("\n");
+            }
+        }
     }
 
     // ------GETTERS------
